requirements.txt
Django==3.2.7
Pillow==8.3.2
PyJWT==2.1.0
asgiref==3.4.1
autopep8==1.5.7
django-cors-headers==3.8.0
djangorestframework-simplejwt==4.8.0
djangorestframework==3.12.4
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
